[drag-spec]: Fix item view state after exiting folder
When the apps grid view detects a drag into one item, if it is not a
folder, we set the state as UI_STATE_DROPPING_IN_FOLDER to give visual
affordance of an attempted folder. If the user drags the item outside
of the attempted folder, the folder should return to either
UI_STATE_NORMAL with the correct scale depending on the state it should
be handling if the apps grid view is in cardified state.

Bug: 1114817
Change-Id: I06ed8064bc83278b9972e2480af5a0746887f420
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2757834
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#863906}