[CrOS Cellular] Track reason for inhibiting cellular scans
We are updating the settings to disable functionality while the Cellular
Device is inhibited. As part of this UI, we now need to show a string
which tells the user why functionality cannot be used.

This CL updates the CellularInhibitor class to provide information about
the reason why it is inhibited, and additionally adds an observer
interface that can be used to detect when the inhibit state has changed.

Bug: 1093185
Change-Id: I1dae4553bb88156f210d475c33dc1c8c0357cd69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2766123
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Azeem Arshad <azeemarshad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#863910}