Add NetworkServiceSandboxLPAC feature to enable LPAC for network process.
The LPAC sandbox has a few advantages over the restricted token sandbox
 - no filesystem intercepts are needed.
 - SSPI works out of the box.
 - simplicity and it's supported by MS.

Unfortunately because of how LPAC works, the major drawback is that LPAC
blocks access to localhost and this cannot be solved without adding
a loopback exemption which requires Admin access.

This issue will have to be resolved before shipping this sandbox.

BUG=841001

Change-Id: I97df5a19577759aefb809aaf89eb2f21cf3904ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2724953
Reviewed-by: James Forshaw <forshaw@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#863911}