Windows Spellcheck: Renderer process can hang retrieving suggestions
If there are a large number of misspellings in editable content (try
some 2000 characters of Lorem Ipsum) and you right click on a
misspelled word to bring up the context menu with suggested
corrections, the renderer process will stop responding if Windows
platform spellchecking is being used.

The main issue is that if suggestions have not already been retrieved
and cached, a synchronous IPC is made from the renderer process to the
browser process, which will not return until GetPerLanguageSuggestions
completes. If the platform spellchecker is still busy performing a text
check on all of the content, this can be a long wait.

A reason that the full text check may take a long time is that
suggestions are being retrieved for all misspelled words, even though
the user may never invoke the context menu to view them.

The fix is to postpone the retrieval of platform spellchecker
suggestions until SpellingMenuObserver::InitMenu is called on the
browser process side. This is done asynchronously following the same
pattern already used for the remote enhanced spellcheck service.
Placeholder menu items are added to the context menu, with one of them
presenting an animation while the retrieval is underway. The callback
when the retrieval completes updates the menu items with a combination
of suggestions from Hunspell (still retrieved on the renderer side)
and the platform.

Other fixes:
1) The renderer side method ContextMenuController::ShowContextMenu was
   retrieving suggestions even for words not marked as misspelled,
   passing an empty string even for Hunspell.
2) When the platform spellchecker performs a full text check in method
   RequestTextCheckForAllLanguages, it was retrieving suggestions
   as well. Now suggestions are only retrieved on demand. Since there
   is some perf trade off between pre-caching suggestions and
   retrieving them when the context menu is invoked, this part of the
   fix is placed behind a feature flag:
   WinRetrieveSuggestionsOnlyOnDemand (on by default)

Bug: 1150610
Change-Id: I5e9b69fdc50241553ae91bc4d92052662eaaa008
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2705781
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Bruce Long <brlong@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#863914}