[GPU] Add sequence wait on dependency metric for gpu_scheduler
Add a time reference when the first dependency is added and another one
when the last dependency is removed. Use these times to compute the
deltas for how long it took to solve dependency and how long it took
after that to get scheduled.

TEST: Manual
BUG: b/181148082
Change-Id: I757cf798a2fe9cea11618dbd7cfa801f7674bba0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2732532
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Auto-Submit: Lucas Berthou <berlu@chromium.org>
Commit-Queue: Lucas Berthou <berlu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#863933}