[iOS] Initial layout of the app state safe mode agent
This is just the initial layout. The agent doesn't not run any logic for
the moment. The implementation will be done in a follow-up CL.

See https://chromium-review.googlesource.com/c/chromium/src/+/2725763
for an example on how the safe mode app state agent can be implemented.
You can see that the app agent switches the app state and keep the stage
to safe mode when the safe mode is triggered.

The app state itself will not start/stop safe mode once we use the agent.

Bug: 1178809
Change-Id: Ifc7eb9b0faa99b0ef416d5602ffee3846177b1f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2728856
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#863944}