Move ARC related logic from FullRestoreReadHandler to ArcReadhandler.
This CL creates ArcReadHandler to handle the ARC related logic for
FullRestoreReadHandler. (ARC app launching saving logic is quite
complex. If all code is added in FullRestoreReadHandler, it might be
hard to read and maintain. So move ARC related logic to a separate
class)

Move all ARC related logic from FullRestoreReadHandler to the new
class ArcReadHandler.

Add |session_id_to_window_id_| to record the map from the session id to
the restore window id.

When the task is created, set the map from the task id to the restore
window id based on |session_id_to_window_id_|. And
|session_id_to_window_id_| can be cleared.

|window_id_to_app_id_| is used to record the map from the restore window
id to the app id for ARC windows only.

There will be follow up CLs to handle the scenario that the ARC window
is created first before the task is created.

BUG=1146900

Change-Id: I11cb096e23e2b6c9a8d039aef70e4a0ce6e78cbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2759809
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#863950}