Fix hotseat swipe location calculation.
The swipe_start_location of hotseat is actually calculated
from the bounds of shelf widget. This works well most of the
cases, but sometimes it may end up with a location outside of
the display; this happens when the shelf is invisible for some
reasons (the virtual keyboard is shown, or the foreground
window is fullscreen).

To avoid the confusion, I think it's better to simply use
the bottom center of the display always.

Bug: 1188301
Test: tast ui.HotseatAnimation with custom fullscreen windows
Change-Id: I408186cfadea04c709a06ff1079b454067448810
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2770123
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#863994}