Files app, Image Loader, Audio and Video Players: Move base/js/ inside file_manager/common/js/
This is a preparation to be able to run optimize_webui=false loading
each JS module individually.

This is necessary because files in the base/js folder aren't inside any
extension causing them to not be loaded via chrome-extension:// URLs.

NB: This is CL:2673587 by lucmult@ reapplied to the main branch.

Bug: 1030998
Change-Id: I3fc0a0eaf6b5c1a7cf1646cc53b18727711ce5f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2766826
Reviewed-by: Bo Majewski <majewski@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864054}