tracing: Fix data race with the SystemTracingService mojo remote.
The race happens because we bind the mojo::SharedRemote on the IO
thread, and the mojo::SharedRemote::is_bound() is checked on the
Perfetto task runner. An internal pointer is read/assigned on different
threads and leads to a potential race.
Fix the race by binding the remote on the Perfetto task runner when the
pending remote is received.

Bug: 1187069, b:147789115
Test: content_browsertests BackgroundTracingManagerBrowserTest.TraceMetadataInTrace passes without DCHEKC() crash.

Change-Id: Ie8f94df95492839f840d3e50f186f1fe61c6f5ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2766975
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864082}