Add accessibilityPrivate API method to show confirmation dialog.
Adds a method to the Accessibility Private JS extension API to
show a confirmation dialog with a given title and description,
and to call a callback when the dialog is accepted or canceled/closed.

Moves AccessibilityPrivateAPITest into chrome/browser/ash/ so that
it can access AccessibilityController.

Bug: 1103799
Change-Id: I0618b2ef7b5e56b05e769d29b4bf9841ba4179fd
AX-Relnotes: N/A
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2765388
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864092}