Replace all uses of Manifest::Location enum with Mojo type (2/n)
As a pre-work to convert ExtensionMsg_Loaded to Mojo, this CL
replaces uses of Manifest::Location enum type with
extensions::mojom::ManifestLocationType progressly because it's
a too huge work to convert them at once.

As the second step, this CL converts the native type to
Mojo type in Extension::Create() method.

AX-Relnotes: n/a

Bug: 1146100
Change-Id: Ifd84d87a86dbd9014fe6c6979c0d92bb38b8472d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2757592
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Sergey Volk <servolk@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Gyuyoung Kim <gyuyoung@igalia.com>
Cr-Commit-Position: refs/heads/master@{#864093}