Files app: Rename selection model's selected item instead of lead item
When the target of the rename event is the file list (which excludes
directory tree renames), the selection model is used to determine what
file is currently selected, and perform necessary checks to enable or
disable the rename command.

However, in the naming controller, the rename is always intiated on the
lead item, which can be distinct from the selection model's selected
item. This makes it possible, for example, to rename the Downloads
folder, leading to unexpected behaviors.

Fix by renaming the selected item instead of the lead item.

Bug: 1094260
Test: browser_tests --gtest_filter=*fileListRenameSelectedItem
Change-Id: Ib12363d7b01b5007af25767df2bba397cb5d465e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2741036
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864097}