[Extensions] Fix regression in RulesCacheDelegate::UpdateRules.
RulesCacheDelegate::UpdateRules will crash if an update task is
posted and the extension is uninstalled before the task runs. This
fixes the bug, which is a regression of crbug.com/517492.

This CL also removes a workaround that was added to a
declarativeContent API test to fix flaky failures as a result of
this bug.

Bug: 517492, 1013457, 1184472
Change-Id: I769224907c4ec7b44a7d945a612bbeacf893ffdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2760501
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864128}