Add an autotest api for installing borealis
We will be diverging the behaviour of UI-based borealis launches (and
therefore installs) from command-line-based ones (see
go/borealis-uninit). As a pre-step we will be migrating our tast tests
to use the UI-based launch/install, rather than the command-line-based
one.

Crostini made a similar move but chose to interface with the actual UI
elements to perform the install, For this implementation i chose to
implement a delegate to the installer which will be used *instead* of
the UI, this will be less brittle to UI tweaks than crostini's
approach.

Bug: b/171668019
Change-Id: I8f481412b7773445089c5e8d499b0eec6f259af0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2751227
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Daniel Ng <danielng@google.com>
Cr-Commit-Position: refs/heads/master@{#864132}