gurl: Change handling of host names containing invalid escapes and invalid utf-8
If a host name contains an invalid escape, e.g. '%t', gurl replaces
it with an escaped '%', resulting in '%25t'.

If a host name contains a valid escape that leads to invalid UTF-8,
gurl replaces it with 'REPLACEMENT CHARACTER' (U+FFFD) encoded in UTF-8:
't%ae' becomes 't%EF%BF%BD'.

But before this change, if a host name contained both of these, then
gurl would do the first step but not the second, which resulted in a
host name that's not valid UTF-8. With this change, a host name
having both issues receives both fixups: '%T%ae' becomes
'%25t%EF%BF%BD'.

Because KURL tries to UTF-8 decode the output of GURL's parser and
sets KURL::string_ to null if that fails, this change
means string_ in KURL in blink is no longer IsNull() for
urls containing percent escapes that previously led to invalid utf-8.

KURL doesn't seem to follow a very principled design for urls that are
not IsValid() (and the existing checks for IsNull() were mostly
sprinkled in reactively to fix crashes:
https://chromiumcodereview.appspot.com/24095009/), so this changes
the behavior for some kurl unit tests. It looks like these tests
exist to make sure that things don't crash and the actual result
isn't all that important.

With this, both valid and invalid urls should have non-null string_
fields in kurl. This also makes the behavior of converting a string
to a KURL and that to a GURL the same as the behavior of converting
a string to a GURL without the intermediary step.

Bug: 1176510
Change-Id: I47145b167e79633e096457b08a5dc448cfe7ddc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2766342
Auto-Submit: Nico Weber <thakis@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864138}