[ios] Make generation of .pak more robust
The internal build of Chrome on iOS needs to add some additional
resources in the final .pak files but still want to pack all the
resources used upstream.

Previously this was done by having a duplicated set of targets in
the internal repository which was brittle as they had to be kept
in sync with the definition in ios/chrome/app/resources/BUILD.gn.

This CL define a set of gn variables that can be used to list the
additional localisable and/or scalable resources to includes when
creating the final .pak file.

The old way to select a different target to generate the .pak
files is kept until the internal code has been updated to set the
new variables.

Bug: 1188640
Change-Id: I15208599fbfc0086df5c57e3e54f3c6f5231a3bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2767344
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864156}