[signin] Fix crash in PrimaryAccountAccessTokenFetcher
The PrimaryAccountAccessTokenFetcher is designed to support outliving
the IdentityManager. The IdentityManager cancels the pending fetchers
when it is shutdown.

However, due to some edge case on Android, the fetcher sometimes wants
to retry the request at this time:
https://source.chromium.org/chromium/chromium/src/+/master:components/signin/public/identity_manager/primary_account_access_token_fetcher.cc;drc=de8a30286744b0f87ad17a82a91efce317898c2c;l=125

This logic causes the fetcher to call into the identity manager, which
is in the middle of shutting down, and crashes.

It is not possible to fix the problem by simply changing the order of
destruction in IdentityManager::Shutdown, because there is a cycle.
The PrimaryAccountManager depends on the token service, as per:
https://source.chromium.org/chromium/chromium/src/+/master:components/signin/internal/identity_manager/primary_account_manager.h;drc=e6508dabc2f96b1afa98079e532ca435dc9e7a66;l=178
And the callstack in https://crbug.com/1187288 shows the opposite
dependency.

This CL adds full support for identity manager shutdown in the primary
account access token fetcher.

Fixed: 1187288
Change-Id: Ieb21134627dbe5c6aa786d85727179085647bf60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2757332
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864158}