[Web Payment] JIT install when SW is installed
Before this patch, if a service worker is already installed, the payment
handler could not be installed just-in-time, because the code waited to
be notified that the service worker has been installed, but never
received this notification.

This patch checks for the presence of the service worker first. If the
service worker is already present, then the payment handler install
proceeds.

After this patch, a payment handler can be installed just in time even
if a service worker is already installed.

Bug: 1154052
Change-Id: Iea3b3c8b845548ee89386dcb59d889eb1ef0bce6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2765465
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Sahel Sharify <sahel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864230}