[Autofill] Preselect country when creating an new autofill profile
The first entry in the country dropdown box is what we believe the
user's country is. It is inferred from Chrome's environment. It is
possible to leave the country empty, but that doesn't guarantee
that autofill won't guess and set the value without user input.

We now preselect the first value.

It is still allowed to leave the country empty. Autofill may set it
implicitly when saving.

Bug: 100845
Change-Id: Ide0acbe0bf7cdebdc2ac8c12c4d8e8638ec0d9b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2764452
Commit-Queue: Christos Froussios <cfroussios@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864249}